Add aloowed_targets
This adds an option for a scope to restrict what target accounts can
be selected.  In practice, this isn't really an issue since users
tend to use the scope created for their account.  But having this option
gives additional control.

Updates to python:3 broke the docker build so it now explicitly specifies
3.8-buster.